+++ src/shared/ui/postLikesBox/PostLikesBox.tsx
@@ -2,6 +2,7 @@
 
 import { ComponentPropsWithoutRef, useMemo } from 'react'
 
+import { usePostLikes } from '@/src/shared/hooks/usePostLikes'
 import { useGetPostLikesQuery } from '@/src/shared/model/api/postsApi'
 import { AvatarBox } from '@/src/shared/ui/avatar/AvatarBox'
 import { Typography } from '@/src/shared/ui/typography/Typography'
@@ -14,24 +15,33 @@ type Props = {
   postId: number
 } & ComponentPropsWithoutRef<'div'>
 
-export const PostLikesBox = ({ className, likesCount, postId }: Props) => {
-  const { data: likesData } = useGetPostLikesQuery(
-    {
-      pageSize: 3,
-      postId,
-    },
-    { refetchOnMountOrArgChange: true }
-  )
+export const PostLikesBox = ({ className, postId }: Props) => {
+  // const {
+  //   data: likesData,
+  //   isFetching,
+  //   isLoading,
+  // } = useGetPostLikesQuery(
+  //   {
+  //     pageSize: 3,
+  //     postId,
+  //   },
+  //   { refetchOnMountOrArgChange: true }
+  // )
+  //
+  // const likesCount = likesData?.totalCount
+  // const isLikesLoaded = !isLoading && !isFetching
+
+  const { avatars, likesCount } = usePostLikes(postId)
 
   const avatarClasses = [s.firstAvaLike, s.secondAvaLike, s.thirdAvaLike]
 
-  const avatars = useMemo(() => {
-    return [...(likesData?.items || [])]
-      .sort((a, b) => b.createdAt.localeCompare(a.createdAt))
-      .map(user => user.avatars?.[0]?.url)
-      .filter(Boolean)
-      .slice(0, 3)
-  }, [likesData])
+  // const avatars = useMemo(() => {
+  //   return [...(likesData?.items || [])]
+  //     .sort((a, b) => b.createdAt.localeCompare(a.createdAt))
+  //     .map(user => user.avatars?.[0]?.url)
+  //     .filter(Boolean)
+  //     .slice(0, 3)
+  // }, [likesData])
 
   return (
     <div className={clsx(s.postLikes, className)}>
